Update linked_list_nth_to_last.py
I changed the method #1 of function print_nth_from_last. I removed the print cur.data and added it to return. Then for method #2, I added a condition  of "count < n" on "if not q", so that we can return the result if the argument is "print(llist.print_nth_from_last(4))". Thank you.